Mass beautify for code

+++ src/modules/IdeaList/components/Idea/index.js
@@ -5,50 +5,41 @@ import React, { Component, PropTypes } from 'react';
 class Idea extends Component {
 
   onClickRemove = () => {
+    if (!confirm("Are you sure you want to delete " + this.props.text + " ?")) {
+      return;
+    }
     const { onRemove } = this.props;
     const key = this.props.id;
     onRemove(key);
-    // this.props.onGetIdeaList();
-  }
-
-  // onClickLike = () => {
-  //   const { onLike } = this.props;
-  //   const key = this.props.id;
-  //   onLike(key, this.props.liked);
-  //   // this.props.onGetIdeaList();
-  // }
+  };
 
   onClickUpdateItem = () => {
     const key = this.props.id;
     const { onUpdateItem } = this.props;
     const text = prompt("Edit task", this.props.text);
-    this.props.onUpdateItem(key, text);
-    // this.props.onGetIdeaList();
-  }
+    onUpdateItem(key, text);
+  };
 
 
   render() {
-    const{
+    const {
       onClickRemove,
-      // onClickLike,
       onClickUpdateItem,
-      props : {
-        likes,
-        text,
-        id,
+      props: {
         editable,
-    }} = this;
+      }
+    } = this;
 
     return (
       <div>
         <div className="idea-item">
           <span className="like-mark">&#9899; {this.props.likes}</span>
           <span>{this.props.text}</span>
-          { editable &&
-            <button className="x-mark" onClick={onClickRemove}>REMOVE</button>
+          {editable &&
+          <button className="x-mark" onClick={onClickRemove}>REMOVE</button>
           }
-          { editable &&
-            <button onClick={onClickUpdateItem} className="edit-mark">EDIT</button>
+          {editable &&
+          <button onClick={onClickUpdateItem} className="edit-mark">EDIT</button>
           }
         </div>
       </div>
@@ -58,20 +49,17 @@ class Idea extends Component {
 
 const mapStateToProps = state => ({
   ideaList: state.idea.ideaList,
-})
+});
 
 const mapDispatchToProps = dispatch => ({
-  onUpdateItem(key, newText){
-    dispatch(update(key, {text: newText}));
+  onUpdateItem(key, newText) {
+    dispatch(update(key, { text: newText }));
   },
 
-  onRemove(key){
+  onRemove(key) {
     dispatch(remove(key));
   },
-  // onLike(key, isLiked){
-  //   dispatch(update(key, {liked: !isLiked}));
-  // },
-  onGetIdeaList(){
+  onGetIdeaList() {
     dispatch(getIdeaList());
   },
   dispatch,
@@ -83,7 +71,6 @@ Idea.propTypes = {
   onRemove: PropTypes.func.isRequired,
   editable: PropTypes.bool.isRequired,
   onUpdateItem: PropTypes.func.isRequired,
-  // onLike: PropTypes.func.isRequired,
 };
 
 export default connect(mapStateToProps, mapDispatchToProps)(Idea);
